# apps/docs/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mayarin API",
    "version": "2026-08-11",
    "description": "Create fiat-denominated commerce requests, accept supported crypto assets, and settle merchants in stablecoins. Versioned operations live under `/v1`; buyer-facing checkout, invoice, and health routes are intentionally unversioned."
  },
  "servers": [
    {
      "url": "https://api.mayarin.xyz",
      "description": "Mainnet"
    },
    {
      "url": "https://api-testnet.mayarin.xyz",
      "description": "Testnet"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "secretKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-side `sk_` key. Grants a subset of the merchant's permissions."
      },
      "publishableOrSecretKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Browser-safe `pk_` key, or a stronger server-side `sk_` key. Identifies the merchant for catalog read and cart checkout; carries no permissions."
      }
    },
    "schemas": {
...
      "Money": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "description": "Exact minor units. Calculate with this value."
          },
          "asset": {
            "type": "string",
            "description": "ISO asset code, e.g. IDR, USD, USDC."
          },
          "formatted": {
            "type": "string",
            "description": "Machine decimal; ungrouped and dot-separated."
          },
          "display": {
            "type": "string",
            "description": "Localized display text. Never parse this value."
          }
        },
        "required": ["amount", "asset", "formatted", "display"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": ["code", "message", "retryable"]
          }
        },
        "required": ["error"]
      },
...
      "X402Resource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "listed": {
            "type": "boolean",
            "description": "Whether it appears in the public x402 resource index."
          },
          "accepts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "chain": {
                  "type": "string"
                },
                "asset": {
                  "type": "string"
                },
                "contract": {
                  "type": "string"
                },
                "payTo": {
                  "type": "string"
                },
                "transferMethod": {
                  "type": "string"
                },
                "domain": {
                  "type": "object",
                  "additionalProperties": {}
                }
              },
              "required": ["chain", "asset", "contract", "payTo", "transferMethod", "domain"]
            }
          }
        },
        "required": ["id", "merchantId", "url", "price", "maxTimeoutSeconds", "listed", "accepts"]
      }
    },
...
  },
  "paths": {
...
    "/v1/x402/resources/{id}/list": {
      "post": {
        "tags": ["x402 resources"],
        "operationId": "listX402Resource",
        "summary": "List a resource in the public index",
        "description": "Opts the resource into the public cross-merchant x402 index. The resource stays sellable by URL either way; listing only decides whether an agent browsing the index is shown it.",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Listed resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/X402Resource"
                }
              }
            }
          },
          "404": {
            "description": "No resource with that id, or not this merchant's.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/x402/resources/{id}/unlist": {
      "post": {
        "tags": ["x402 resources"],
        "operationId": "unlistX402Resource",
        "summary": "Remove a resource from the public index",
        "security": [
          {
            "secretKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Unlisted resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/X402Resource"
                }
              }
            }
          },
          "404": {
            "description": "No resource with that id, or not this merchant's.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}